pyproject.toml
[project]
name = "avanamy-backend"
version = "0.1.0"
description = "Backend for Avanamy API change detection"
authors = [
    { name = "Ken Gavazzi", email = "kenneth.gavazzi@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2,<0.122.0",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "boto3 (>=1.40.76,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.17.2,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "opentelemetry-sdk (>=1.38.0,<2.0.0)",
    "opentelemetry-api (>=1.38.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.59b0,<0.60)",
    "markdown (>=3.10,<4.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "anthropic (>=0.75.0,<0.76.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0,<7.0)",
    "clerk-backend-api (>=4.2.0,<5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
]

# ---------------------------
# Poetry-specific settings
# ---------------------------

[tool.poetry]
packages = [
    { include = "avanamy", from = "src" }
]

[tool.poetry.scripts]
# Console entry removed: CLI moved/removed. If you need a console script,
# re-add an entry here pointing to the desired callable.

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)"
]

[tool.pytest.ini_options]
markers = [
    "anyio: mark test to run with anyio",
]
# Note: anyio will try to run tests with both asyncio and trio backends.
# Since trio is not installed, use -k "not trio" to exclude trio tests.
# Example: pytest -k "not trio" -v
